Extract manual exposure policy into camera.policy
Pure refactor; behavior preserved from flat camera package.
